report bug fix

+++ src/main/java/Internal/framework/ui/ApplicationFrm.java
@@ -287,10 +287,10 @@ public class ApplicationFrm extends javax.swing.JFrame{
             wd.show();
 
             // compute new amount
-            long deposit = Long.parseLong(amountDeposit);
+            double deposit = Double.parseDouble(amountDeposit);
             String samount = (String)model.getValueAt(selection, 5);
-            long currentamount = Long.parseLong(samount);
-            long newamount=currentamount-deposit;
+            double currentamount = Double.parseDouble(samount);
+            Double newamount=currentamount-deposit;
             model.setValueAt(String.valueOf(newamount),selection, 5);
             if (newamount <0){
                 JOptionPane.showMessageDialog(JButton_Withdraw, " Account "+accnr+" : balance is negative: $"+String.valueOf(newamount)+" !","Warning: negative balance",JOptionPane.WARNING_MESSAGE);
